feat: #571 technology report - move componente TimeTange to shared
