+++ src/app/modules/technology-report/pages/technology-report.component.html
@@ -1,1 +1,2 @@
+<app-time-range-form></app-time-range-form>
 <app-technology-report-table></app-technology-report-table>
